+++ arealite/engine/vl_fsdp_engine.py
@@ -225,4 +225,3 @@ class VL_FSDPEngine(FSDPEngine):
         mb_list = unsqueeze_mb_list(mb_list)
         return mb_list
 
-
